changed everything from double to float

+++ Scene.cpp
@@ -56,10 +56,10 @@ Scene::~Scene() {
 	}
 }
  
-Ray Scene::castEyeRay(double i, double j){
-	double alpha = (2.0*i-width)/width;
+Ray Scene::castEyeRay(float i, float j){
+	float alpha = (2.0*i-width)/width;
 	alpha *=tan(fovx/2.0);
-	double beta = (2.0*j-height)/height;
+	float beta = (2.0*j-height)/height;
 	beta *= tan(fovy/2.0);
 	
 	Ray ray(eye, alpha*u + beta*v - w);
@@ -100,7 +100,7 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 	} else if (cmd == "output") {
 		line >> filename;
 	} else if (cmd == "camera") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1 >> arg2 >> arg3;
 		eye = vec3(arg1,arg2,arg3);
 		
@@ -112,10 +112,10 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 		setCoordinateFrame(lookat,up);
 		line >> fovy;
 		fovy*=pi/180.0;
-		double d = height/(2.0 * tan(fovy*0.5));
+		float d = height/(2.0 * tan(fovy*0.5));
 		fovx = 2. * atan(width/(2.0*d));
 	} else if (cmd == "sphere") {
-		double arg1, arg2, arg3, arg4;
+		float arg1, arg2, arg3, arg4;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
@@ -149,14 +149,14 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 		verts.reserve(maxvertnorms);
 		verts.reserve(maxvertnorms);
 	} else if (cmd == "vertex") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
 		vec3 v(arg1,arg2,arg3);
 		verts.push_back(v);
 	} else if (cmd == "vertexnormal") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
@@ -207,20 +207,20 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 		t->refractivity = refractivity;
 		objects.push_back(t);
 	} else if(cmd == "translate") {
-		double arg1,arg2,arg3;
+		float arg1,arg2,arg3;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
 		mv.top() *= Transform::translate(arg1, arg2, arg3);
 	} else if(cmd == "rotate") {
-		double arg1,arg2,arg3,arg4;
+		float arg1,arg2,arg3,arg4;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
 		line >> arg4;
 		mv.top() *= Transform::rotate(arg4,vec3(arg1,arg2,arg3));
 	} else if (cmd=="scale") {
-		double arg1,arg2,arg3;
+		float arg1,arg2,arg3;
 		line >> arg1;
 		line >> arg2;
 		line >> arg3;
@@ -230,15 +230,15 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 	} else if (cmd == "popTransform"){
 		mv.pop();
 	} else if (cmd == "directional") {
-		double x,y,z,r,g,b;
+		float x,y,z,r,g,b;
 		line >> x >> y >> z >> r >> g >> b;
 		vec3 dir = vec3(mv.top()*vec4(x,y,z,0.0));
 		DirectionalLight* light = new DirectionalLight(vec3(r,g,b),dir);
 		lights.push_back(light);
 	} else if (cmd == "point") {
-		double x,y,z,r,g,b;
+		float x,y,z,r,g,b;
 		line >> x >> y >> z >> r >> g >> b;
-		vec3 point = vec3(mv.top()*vec4(x,y,z,1.0));
+		vec3 point = vec3(mv.top()*vec4(x,y,z,1.0f));
 		PointLight* light = new PointLight(vec3(r,g,b),vec3(x,y,z),constant,linear,quadratic);
 		light->shadowrays = shadowrays;
 		light->lightradius = lightradius;
@@ -246,21 +246,21 @@ void Scene::parseLine(string l, stack<mat4>& mv, vector<vec3>& verts,
 	} else if (cmd == "attenuation") {
 		line >> constant >> linear >> quadratic;
 	} else if (cmd == "ambient") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1 >> arg2 >> arg3;
 		ambient = vec3(arg1,arg2,arg3);
 	} else if (cmd == "diffuse") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1 >> arg2 >> arg3;
 		diffuse = vec3(arg1,arg2,arg3);
 	} else if (cmd == "specular") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1 >> arg2 >> arg3;
 		specular = vec3(arg1,arg2,arg3);
 	} else if (cmd == "shininess") {
 		line >> shininess;
 	} else if (cmd == "emission") {
-		double arg1, arg2, arg3;
+		float arg1, arg2, arg3;
 		line >> arg1 >> arg2 >> arg3;
 		emission = vec3(arg1,arg2,arg3);
 	} else if (cmd == "indexofrefraction") {
