+++ Transform.cpp
@@ -3,12 +3,12 @@
 #include "Transform.h"
 #include <iostream>
 
-mat4 Transform::rotate(double degrees, const vec3& axis) {
+mat4 Transform::rotate(float degrees, const vec3& axis) {
 	mat3 parallel( cos(degrees*pi/180.0) );
 	vec3 a = glm::normalize(axis);
-	double x = a[0];
-	double y = a[1];
-	double z = a[2];
+	float x = a[0];
+	float y = a[1];
+	float z = a[2];
 	mat3 rotation(x*x, x*y, x*z, x*y, y*y, y*z, x*z, y*z, z*z);
 	rotation *= (1-cos(degrees*pi/180.0));
 	mat3 cross(0.0, -z, y, z, 0.0, -x, -y, x, 0.0);
@@ -18,12 +18,12 @@ mat4 Transform::rotate(double degrees, const vec3& axis) {
 	return ret;
 }
 
-mat4 Transform::scale(double sx, double sy, double sz) {
+mat4 Transform::scale(float sx, float sy, float sz) {
 	mat4 S(sx,0,0,0,0,sy,0,0,0,0,sz,0,0,0,0,1);
 	return glm::transpose(S);	
 }
 
-mat4 Transform::translate(double tx, double ty, double tz) {
+mat4 Transform::translate(float tx, float ty, float tz) {
 	mat4 T(1,0,0,tx,0,1,0,ty,0,0,1,tz,0,0,0,1);
 	return glm::transpose(T);	
 }
